Brings the warband phase1..phase9 chain into the integration assembly. The chain
is strictly linear, so this single merge lands all nine phases.

The reason this could not wait: integration carried the ORIGINAL, unhardened
CMSG_TRANSFER_CURRENCY_FROM_ACCOUNT_CHARACTER handler while its owning branch had
already fixed it. Integration's copy read the source balance from an async query
snapshot, checked it, then wrote an ABSOLUTE `SET Quantity = ?`, with no
in-flight guard - so N spammed packets all read the same pre-transfer balance,
all passed the check, all wrote the same absolute value (source debited once) and
all credited the destination. A live currency duplication. The owner's version -
now the merged one - takes an atomic sWorld->BeginCurrencyTransfer reservation,
deducts with a guarded `SET Quantity = Quantity - ? WHERE ... AND Quantity >= ?`,
and orders the debit before the credit so a crash loses currency rather than
duplicating it.

Conflict resolution (33 files, ~100 hunks) was a union merge: both parents are
our own work and each added different systems to the same files. Notable:

* CHAR_INS_CHARACTER: both parents added columns to the same positional INSERT.
  Merged to 87 columns/87 placeholders ending
  "... chromieTimeExpansionId, timerunningSeasonId, battlenetAccount", with
  Player::SaveToDB binding in that exact order. A mismatch here corrupts
  character saves silently.
* CHAR_INS_WARBAND_CURRENCY_TRANSFER_LOG went 6 -> 7 columns (receivedQuantity);
  call site, SELECT and DDL verified consistent.
* CONDITION enum: both parents used 60 for different enumerators. Kept ours
  (CHROMIE_TIME=60, GROUP_STATUS=61, COVENANT=62) and renumbered the two
  conditions rows in the owner's 2026_07_16_00_world.sql from 60 to 61, which
  would otherwise have silently mis-targeted gossip menu 1541.
* WarbandScenePlacement DB2: kept ours - our own 2026_07_21_01_hotfixes.sql
  documents that layout EF845637 is what 12.0.7.68275 ships and that the older
  form aborts LoadDB2 at startup.
* Dropped the owner's single-garrison `_garrison` form: that member no longer
  exists (Player.h carries the `_garrisons` map), so it would not compile.

Also repaired several same-content blocks that git auto-merged at different
offsets into redefinitions (a duplicated ~40-entry warband enum, duplicate
PlayerCompanionInfo statements/stores/structs, a duplicated LoginQueryHolder
block). Those are outside the conflict hunks and would each have broken the build.

Verified: a guard recording all 948 identifiers unique to either parent reports
every one still present; no conflict markers; no duplicate enumerators,
prepared-statement ids, DEFINE_HANDLERs, DB2 stores or LoadInfos introduced.

NOT BUILT. A compile is still the real check before this goes near a realm.

+++ src/server/game/Entities/Player/SceneMgr.cpp
@@ -59,7 +59,7 @@ uint32 SceneMgr::PlaySceneByTemplate(SceneTemplate const* sceneTemplate, Positio
 
     WorldPackets::Scenes::PlayScene playScene;
     playScene.SceneID              = sceneTemplate->SceneId;
-    playScene.PlaybackFlags        = sceneTemplate->PlaybackFlags.AsUnderlyingType();
+    playScene.PlaybackFlags        = int32(sceneTemplate->PlaybackFlags.AsUnderlyingType());
     playScene.SceneInstanceID      = sceneInstanceID;
     playScene.SceneScriptPackageID = sceneTemplate->ScenePackageId;
     playScene.Location             = *position;
